backend/requirements.txt
Flask>=3.0.0
waitress>=2.1.0
face-recognition>=1.3.0
Pillow>=10.0.0
haversine>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0

#dev
pytest>=8.0.0
ruff>=0.6.0
black>=24.0.0